// apps/web/packages/holand-core/src/ui/floating/types.ts
export type FloatingPlacement = 'end' | 'start';
export type FloatingSide = 'top' | 'right' | 'bottom' | 'left';
export type FloatingPosition =
  | FloatingSide
  | `${FloatingSide}-${FloatingPlacement}`;

export type TextDirection = 'ltr' | 'rtl';

export type TooltipPreset =
  | 'toolbar'
  | 'header-edge'
  | 'media'
  | 'sidebar';